Managed to print out multiple images on screen without instability
Canvas.Java
Changed from System.out.print to BufferedWriter as it is supposed to be faster
Corrected screenSize
Wrote a fullPrint first which reprints the whole screen. This proved to be slow,
so a print that only updates the required pixels was added.

Crawler.Java
Now main prints two images, toggling between them.