Added an ugly little window which takes key inputs.
Apparently Java doesn't have any real way of just reading key presses.
The most obvious way is to use a JFrame window, which can catch these presses IF the window is selected.
Of course the idea is that you're also supposed to put the game in that window. But who can be bothered.

Removed classes that are not used.
Added Keysniffer class, called magically by the JFrame frame.
Now moves around tile by tile.
Added a block so you can't move off the map.